导入导出接口、模拟多线程首页

+++ module/src/main/java/red/mlz/module/module/channel/entity/Channel.java
@@ -18,5 +18,5 @@ public class Channel {
     //更新时间
     private Integer updatedTime;
     //删除
-    private BigInteger isDeleted;
+    private Integer isDeleted;
 }
