+++ module/src/main/java/red/mlz/module/module/event/entity/Event.java
@@ -3,12 +3,14 @@ package red.mlz.module.module.event.entity;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigInteger;
+
 
 @Data
 @Accessors(chain = true)
 public class Event {
     //
-    private Long id;
+    private BigInteger id;
     //活动内容
     private String content;
     //创建时间
@@ -16,5 +18,5 @@ public class Event {
     //更新时间
     private Integer updatedTime;
     //删除
-    private Byte isDeleted;
+    private Integer isDeleted;
 }
